Change internal integer representation of Value node
A Value node would store an integer as a long.  This causes needless
portability risks, as long can be of varying sizes.  Change it to use
int instead.  All code using this was already careful to only store
32-bit values anyway.

+++ src/backend/nodes/value.c
@@ -20,7 +20,7 @@
  *	makeInteger
  */
 Value *
-makeInteger(long i)
+makeInteger(int i)
 {
 	Value	   *v = makeNode(Value);
